move getResolvedDirOrFile for peers to its own file
